Add FastAPI app setup with Dockerfile and CI/CD workflow configuration

+++ Reema Alhenaki/FastAPI_app/app/crud.py
@@ -1,39 +1,16 @@
 from sqlalchemy.orm import Session
 from app.models import HISPatient
 from app.schemas import PatientCreate, PatientUpdate
-import random
 from datetime import datetime
 from fastapi import HTTPException
+import app.models
 
-def generate_unique_patient_id():
-    return random.randint(100000, 999999)
 
 
 def create_patient(db: Session, patient: PatientCreate):
-    # Check for required fields 
-    required_fields = {
-        "FirstName": patient.FirstName,
-        "LastName": patient.LastName,
-        "Gender": patient.Gender,
-        "DateofBirth": patient.DateofBirth,
-        "NationalityID": patient.NationalityID,
-    } 
-
-    missing = [field for field, value in required_fields.items() if value is None]
-    if missing:
-        raise HTTPException(
-            status_code=422,
-            detail=f"The following required fields are missing or null: {', '.join(missing)}"
-        )
-
-    new_id = generate_unique_patient_id()
-    while db.query(HISPatient).filter(HISPatient.PatientID == new_id).first():
-        new_id = generate_unique_patient_id()
-
+    
     now = datetime.now()
-
     db_patient = HISPatient(
-        PatientID=new_id,
         FirstName=patient.FirstName,
         MiddleName=patient.MiddleName,
         LastName=patient.LastName,
@@ -44,14 +21,12 @@ def create_patient(db: Session, patient: PatientCreate):
         FirstVisit=now,
         LastVisit=now,
         NoOfVisit=0,
-        MobileNumber=int(patient.MobileNumber) if patient.MobileNumber else None,
+        MobileNumber= patient.MobileNumber if patient.MobileNumber else None,
     )
-
     db.add(db_patient)
     db.commit()
     db.refresh(db_patient)
     return db_patient
-
 
 
 
@@ -62,6 +37,8 @@ def get_patients(db: Session):
 def get_patient(db: Session, patient_id: int):
     return db.query(HISPatient).filter(HISPatient.PatientID == patient_id).first()
 
+
+
 def update_patient(db: Session, patient_id: int, updates: PatientUpdate):
     patient = get_patient(db, patient_id)
     if not patient:
@@ -69,17 +46,28 @@ def update_patient(db: Session, patient_id: int, updates: PatientUpdate):
     update_data = updates.model_dump(exclude_unset=True)
     update_data = {k: v for k, v in update_data.items() if v is not None}
 
-
     for key, value in update_data.items():
         setattr(patient, key, value)
     db.commit()
     db.refresh(patient)
     return patient
 
+
+
 def delete_patient(db: Session, patient_id: int):
     patient = get_patient(db, patient_id)
     if not patient:
         return None
+
+    # Check if patient has appointments
+    appointments = db.query(app.models.t_HIS_Appointment).filter(
+        app.models.t_HIS_Appointment.c.PatientID == patient_id
+    ).first()
+
+    if appointments:
+        raise HTTPException(status_code=400, detail="Cannot delete: patient has appointments.")
+
+    # If no appointments, delete
     db.delete(patient)
     db.commit()
     return patient
